OpenAL efx fixes for macOS/Linux
