+++ Engine/source/sfx/openal/LoadOAL.h
@@ -27,18 +27,10 @@
 #  include "platform/platform.h"
 #endif
 
-#if defined(TORQUE_OS_MAC)
-#  include <OpenAL/al.h>
-#  include <OpenAL/alc.h>
-#elif defined(TORQUE_OS_LINUX)
-#  include <AL/al.h>
-#  include <AL/alc.h>
-#else
-#  include <al/al.h>
-#  include <al/alc.h>
-#  include <AL/alext.h>
-#  include <AL/efx-presets.h>
-#endif
+#include <AL/al.h>
+#include <AL/alc.h>
+#include <AL/alext.h>
+#include <AL/efx-presets.h>
 
 #ifndef ALAPIENTRY
 #define ALAPIENTRY
